requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
click==8.3.1
Django==5.2
django-debug-toolbar
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.10
Faker==40.1.0
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
isort==6.0.1
Markdown==3.8
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.1.1
uvicorn==0.37.0
whitenoise==6.11.0
